# [ROCm] Generalize FastGeLU