--- a/onnxruntime/contrib_ops/rocm/bert/elementwise_impl/impl_gelu.cu
+++ b/onnxruntime/contrib_ops/rocm/bert/elementwise_impl/impl_gelu.cu
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#include "contrib_ops/rocm/bert/elementwise_impl/impl.cuh"
+
+ELEMENTWISE_KERNEL_IMPL(functor::GeLU, float);
+ELEMENTWISE_KERNEL_IMPL(functor::GeLU, half);
+ELEMENTWISE_KERNEL_IMPL(functor::GeLU, BFloat16);