--- a/onnxruntime/contrib_ops/rocm/bert/fast_gelu.h
+++ b/onnxruntime/contrib_ops/rocm/bert/fast_gelu.h
@@ -5,7 +5,6 @@
 
 #include "core/common/common.h"
 #include "core/providers/rocm/rocm_kernel.h"
-#include "contrib_ops/rocm/bert/fast_gelu_impl.h"
 
 namespace onnxruntime {
 namespace contrib {