Allow the whole pipeline to be parameterized with unary elementwise functor.